moved decode_ptr functions to another file
