+++ go.mod
@@ -3,7 +3,7 @@ module github.com/ChainSafe/gossamer
 replace github.com/go-interpreter/wagon v0.0.0 => github.com/perlin-network/wagon v0.3.1-0.20180825141017-f8cb99b55a39
 
 require (
-	github.com/ChainSafe/go-schnorrkel v0.0.0-20191230221737-ebc1c03bedc6
+	github.com/ChainSafe/go-schnorrkel v0.0.0-20191119190741-a5524ad9568e
 	github.com/ChainSafe/log15 v1.0.0
 	github.com/OneOfOne/xxhash v1.2.5
 	github.com/btcsuite/btcutil v0.0.0-20190425235716-9e5f4b9a998d
